chabito-telegram/requirements.txt
python-telegram-bot>=21.0,<22
websockets>=12.0,<14
python-dotenv>=1.0.0,<2

